add gastosDB obtenerGastos

+++ Domain/Dtos/Cliente/Gastos.cs
@@ -9,8 +9,8 @@ namespace Domain.Dtos.Cliente
     public class Gastos
     {
         public string Id { get; set; }
-        public float Valor { get; set; }
-        public int Consumidor { get; set; }
+        public decimal Valor { get; set; }
+        public Int32 Consumidor { get; set; }
         public DateTime Fecha { get; set; }
     }
 }
